# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/edi_documents:
    get:
      tags:
        - EdiDocuments
      summary: List EdiDocuments
      description: Get a paginated list of EdiDocuments with optional filtering and sorting
      operationId: listEdiDocuments
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdiDocumentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - EdiDocuments
      summary: Create EdiDocument
      description: Create a new EdiDocument
      operationId: createEdiDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEdiDocumentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdiDocument'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/trading_partners:
    get:
      tags:
        - TradingPartners
      summary: List TradingPartners
      description: Get a paginated list of TradingPartners with optional filtering and sorting
      operationId: listTradingPartners
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingPartnerList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TradingPartners
      summary: Create TradingPartner
      description: Create a new TradingPartner
      operationId: createTradingPartner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTradingPartnerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingPartner'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    EdiDocument:
      type: object
      required:
        - id
        - company_id
        - partner_id
        - doc_type
        - direction
        - control_number
        - business_key
        - status
        - payload
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        doc_type:
          type: string
        direction:
          type: string
        control_number:
          type: string
        business_key:
          type: string
        status:
          type: string
        payload:
          type: string
        mapped_ref_type:
          nullable: true
          type: string
        mapped_ref_id:
          nullable: true
          type: string
          format: uuid
        error_detail:
          nullable: true
          type: string
        acknowledged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateEdiDocumentRequest:
      type: object
      required:
        - company_id
        - partner_id
        - doc_type
        - direction
        - control_number
        - business_key
        - status
        - payload
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        partner_id:
          type: string
          format: uuid
        doc_type:
          type: string
        direction:
          type: string
        control_number:
          type: string
        business_key:
          type: string
        status:
          type: string
        payload:
          type: string
        mapped_ref_type:
          nullable: true
          type: string
        mapped_ref_id:
          nullable: true
          type: string
          format: uuid
        error_detail:
          nullable: true
          type: string
        acknowledged_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    EdiDocumentList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EdiDocument'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TradingPartner:
      type: object
      required:
        - id
        - company_id
        - name
        - partner_code
        - format
        - partner_direction
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
        partner_code:
          type: string
        format:
          type: string
        partner_direction:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    CreateTradingPartnerRequest:
      type: object
      required:
        - company_id
        - name
        - partner_code
        - format
        - partner_direction
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        name:
          type: string
        partner_code:
          type: string
        format:
          type: string
        partner_direction:
          type: string
        status:
          type: string
        metadata:
          type: object
...
    TradingPartnerList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TradingPartner'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []