# src/exports/types.rs
//! Public types for Edi module
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! These DTOs are the ONLY types other modules should use.
//! They are decoupled from internal domain entities.

use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};
use crate::domain::entity::*;

// ============================================================================
// EDIDOCUMENT TYPES
// ============================================================================

/// Type-safe ID for EdiDocument
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct EdiDocumentId(pub Uuid);

impl EdiDocumentId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for EdiDocumentId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<EdiDocumentId> for Uuid {
    fn from(id: EdiDocumentId) -> Self {
        id.0
    }
}

/// Data transfer object for EdiDocument
///
/// This is the public representation of EdiDocument for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct EdiDocumentDto {
    pub id: EdiDocumentId,
    pub company_id: Uuid,
    pub partner_id: Uuid,
    pub doc_type: EdiDocType,
    pub direction: EdiDirection,
    pub control_number: String,
    pub business_key: String,
    pub status: EdiStatus,
    pub payload: String,
    pub mapped_ref_type: Option<String>,
    pub mapped_ref_id: Option<Uuid>,
    pub error_detail: Option<String>,
    pub acknowledged_at: Option<DateTime<Utc>>,
    pub metadata: serde_json::Value,
}

/// Summary view of EdiDocument for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct EdiDocumentSummary {
    pub id: EdiDocumentId,
    pub status: EdiStatus,
}

/// Reference to EdiDocument for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct EdiDocumentRef {
    pub id: EdiDocumentId,
}

// ============================================================================
// TRADINGPARTNER TYPES
// ============================================================================

/// Type-safe ID for TradingPartner
///
/// Use this instead of raw Uuid for type safety across modules.
#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
pub struct TradingPartnerId(pub Uuid);

impl TradingPartnerId {
    pub fn new(id: Uuid) -> Self {
        Self(id)
    }

    pub fn into_inner(self) -> Uuid {
        self.0
    }
}

impl From<Uuid> for TradingPartnerId {
    fn from(id: Uuid) -> Self {
        Self(id)
    }
}

impl From<TradingPartnerId> for Uuid {
    fn from(id: TradingPartnerId) -> Self {
        id.0
    }
}

/// Data transfer object for TradingPartner
///
/// This is the public representation of TradingPartner for other modules.
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TradingPartnerDto {
    pub id: TradingPartnerId,
    pub company_id: Uuid,
    pub name: String,
    pub partner_code: String,
    pub format: EdiFormat,
    pub partner_direction: PartnerDirection,
    pub status: TradingPartnerStatus,
    pub metadata: serde_json::Value,
}

/// Summary view of TradingPartner for list displays
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TradingPartnerSummary {
    pub id: TradingPartnerId,
    pub name: String,
    pub status: TradingPartnerStatus,
}

/// Reference to TradingPartner for foreign key relationships
#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct TradingPartnerRef {
    pub id: TradingPartnerId,
}

// ============================================================================
// CUSTOM TYPES
// ============================================================================

// <<< CUSTOM TYPES START >>>
// Add custom public types here
// <<< CUSTOM TYPES END >>>